+++ src/devices/compass.cpp
@@ -11,12 +11,12 @@ Compass::Compass (char* device_file, int slave_address)
 {
   m_interface_port = std::unique_ptr<I2C> (
       new I2C (device_file, slave_address));
-  set_device_id (static_cast<int> (Device_ID::COMPASS));
+  set_device_id (Device_ID::COMPASS);
 }
 Compass::Compass (int slave_address)
 {
   m_interface_port = std::unique_ptr<I2C> (new I2C (slave_address));
-  set_device_id (static_cast<int> (Device_ID::COMPASS));
+  set_device_id (Device_ID::COMPASS);
 }
 int
 Compass::read (unsigned char* buf, int length)
