+++ src/devices/gyroscope.cpp
@@ -9,12 +9,12 @@ Gyroscope::Gyroscope (char* device_file, int slave_address)
 {
   m_interface_port = std::unique_ptr<I2C> (
       new I2C (device_file, slave_address));
-  set_device_id(static_cast<int>(Device_ID::GYROSCOPE));
+  set_device_id (Device_ID::GYROSCOPE);
 }
 Gyroscope::Gyroscope (int slave_address)
 {
   m_interface_port = std::unique_ptr<I2C> (new I2C (slave_address));
-  set_device_id(static_cast<int>(Device_ID::GYROSCOPE));
+  set_device_id (Device_ID::GYROSCOPE);
 }
 int Gyroscope::read(unsigned char* buf, int length){
   return -1;
