+++ src/devices/hygrometer.cpp
@@ -9,12 +9,12 @@ Hygrometer::Hygrometer (char* device_file, int slave_address)
 {
   m_interface_port = std::unique_ptr<I2C> (
       new I2C (device_file, slave_address));
-  set_device_id (static_cast<int> (Device_ID::HYGROMETER));
+  set_device_id (Device_ID::HYGROMETER);
 }
 Hygrometer::Hygrometer (int slave_address)
 {
   m_interface_port = std::unique_ptr<I2C> (new I2C (slave_address));
-  set_device_id(static_cast<int>(Device_ID::HYGROMETER));
+  set_device_id (Device_ID::HYGROMETER);
 }
 int
 Hygrometer::read (unsigned char* buf, int length)
