+++ src/devices/wind_sensor.cpp
@@ -9,12 +9,12 @@ WindSensor::WindSensor (char* device_file, int slave_address)
 {
   m_interface_port = std::unique_ptr<I2C> (
       new I2C (device_file, slave_address));
-  set_device_id(static_cast<int>(Device_ID::WIND_SENSOR));
+  set_device_id (Device_ID::WIND_SENSOR);
 }
 WindSensor::WindSensor (int slave_address)
 {
   m_interface_port = std::unique_ptr<I2C> (new I2C (slave_address));
-  set_device_id(static_cast<int>(Device_ID::WIND_SENSOR));
+  set_device_id (Device_ID::WIND_SENSOR);
 }
 int WindSensor::read(unsigned char* buf, int length){
   return -1;
